feat(e2e): push notifications experience
