+++ e2e/conf/browserstack.conf.js
@@ -22,6 +22,7 @@ const specs = [
 	// './experiences/featured_article/featured_article.test.js'
 	// './experiences/homepage/homepage.test.js'
 	// './experiences/multicard/multicard.test.js'
+	// './experiences/push_notifications/push-notifications.test.js'
 	'./e2e/experiences.test.js'
 ];
