+++ e2e/experiences.test.js
@@ -5,6 +5,7 @@ const facebookTest = require('../experiences/facebook/facebook.test');
 const featuredArticleTest = require('../experiences/featured_article/featured-article.test');
 const homepageTest = require('../experiences/homepage/homepage.test');
 const multicardTest = require('../experiences/multicard/multicard.test');
+const pushNotificationsTest = require('../experiences/push_notifications/push-notifications.test');
 const taboolaTest = require('../experiences/taboola/taboola.test');
 const topArticlesTest = require('../experiences/top_articles/top-articles.test');
 
@@ -14,5 +15,6 @@ describe('facebook experience', facebookTest);
 describe('featured article experience', featuredArticleTest);
 describe('homepage experience', homepageTest);
 describe('multicard experience', multicardTest);
+describe('push notifications experience', pushNotificationsTest);
 describe('taboola experience', taboolaTest);
 describe('top articles experience', topArticlesTest);
